feat: Unified home screen - app drawer as home\n\n- HomeFragment removed, AppDrawerFragment is now the start destination\n- Clock, date, favorites, and app list all in one screen\n- Nav graph simplified: mainFragment removed\n- SettingsFragment: blur/overlay removed for performance\n- Performance: LUT-based wave animation, animateLayoutChanges removed\n- Left-hand index side support\n- Default font changed to sans-serif for CJK compatibility\n- Japanese README added\n- Version: v1.0.3-beta
