+++ app/build.gradle
@@ -23,8 +23,6 @@ android {
         resourceConfigurations += ["en", "ar", "de", "es-rES", "es-rUS", "fr", "he", "hr", "hu", "in", "it", "ja", "nl", "pl", "pt-rBR", "ru-rRU", "sv", "tr", "uk", "zh"]
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
 
-        renderscriptTargetApi 31
-        renderscriptSupportModeEnabled true
     }
 
     buildTypes {
@@ -67,7 +65,5 @@ dependencies {
     //Material dependency
     implementation libs.material
 
-    // BlurView for Android 11 and below compatibility
-    implementation 'com.github.Dimezis:BlurView:version-2.0.3'
 
 }
