+++ app/src/main/java/app/olauncher/data/Prefs.kt
@@ -45,6 +45,7 @@ class Prefs(context: Context) {
     private val HOME_BUTTON_SHOW_RECENTS = "HOME_BUTTON_SHOW_RECENTS"
     private val DEFAULT_WALLPAPER_SET = "DEFAULT_WALLPAPER_SET"
     private val FONT_FAMILY = "FONT_FAMILY"
+    private val INDEX_SIDE = "INDEX_SIDE"
 
     private val APP_NAME_1 = "APP_NAME_1"
     private val APP_NAME_2 = "APP_NAME_2"
@@ -230,8 +231,12 @@ class Prefs(context: Context) {
         set(value) = prefs.edit { putBoolean(DEFAULT_WALLPAPER_SET, value).apply() }
 
     var fontFamily: String
-        get() = prefs.getString(FONT_FAMILY, "sans-serif-thin").toString()
+        get() = prefs.getString(FONT_FAMILY, "sans-serif").toString()
         set(value) = prefs.edit { putString(FONT_FAMILY, value).apply() }
+
+    var indexSide: String
+        get() = prefs.getString(INDEX_SIDE, "right").toString()
+        set(value) = prefs.edit { putString(INDEX_SIDE, value).apply() }
 
     var hiddenApps: MutableSet<String>
         get() = prefs.getStringSet(HIDDEN_APPS, mutableSetOf()) as MutableSet<String>
