--- app/src/main/res/layout/fragment_home.xml
@@ -1,201 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  Niagara Launcher 風 ホーム画面レイアウト v2
-    - ガラスカード背景を全て除去（クリーンな見た目）
-    - 時計はシンプルなテキストのみ
-    - 大きなアプリアイコン
-    - アルファベットインデックス右端に常設
-    - ブランド刻印は極小に
-    - レターポップアップ表示領域を追加
--->
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/mainLayout"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:animateLayoutChanges="true"
-    tools:context=".ui.HomeFragment">
-
-    <!-- ロック用プレースホルダー -->
-    <FrameLayout
-        android:id="@+id/lock"
-        android:layout_width="match_parent"
-        android:layout_height="1dp"
-        android:contentDescription="@string/lock_layout_description" />
-
-    <!-- 最近のアプリ用プレースホルダー -->
-    <FrameLayout
-        android:id="@+id/recents"
-        android:layout_width="match_parent"
-        android:layout_height="1dp"
-        android:contentDescription="@string/recents_layout_description" />
-
-    <!-- ガラスエフェクト背景（画面全体をカバー、より透明感のあるNiagara風に）-->
-    <eightbitlab.com.blurview.BlurView
-        android:id="@+id/blurView"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="#99000000"
-        android:visibility="gone" />
-
-    <!-- メインコンテンツ -->
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical">
-
-        <!-- 上部スペーサー（Niagara: 画面上部に余白） -->
-        <Space
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="0.10" />
-
-        <!-- 日時エリア（Niagara: カードなし、純粋なテキスト） -->
-        <LinearLayout
-            android:id="@+id/dateTimeLayout"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="28dp"
-            android:gravity="start"
-            android:orientation="vertical"
-            android:paddingVertical="4dp"
-            android:visibility="gone"
-            tools:visibility="visible">
-
-            <!-- 時計（Niagara: 細字、大きめ） -->
-            <TextClock
-                android:id="@+id/clock"
-                style="@style/TextDefault"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:fontFamily="sans-serif-thin"
-                android:format12Hour="h:mm"
-                android:format24Hour="H:mm"
-                android:includeFontPadding="false"
-                android:letterSpacing="-0.02"
-                android:textSize="@dimen/time_size"
-                tools:text="02:34" />
-
-            <!-- 日付（Niagara: 小さいサブテキスト） -->
-            <TextView
-                android:id="@+id/date"
-                style="@style/TextDefault"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:alpha="0.7"
-                android:format12Hour="EEE, dd MMM"
-                android:format24Hour="EEE, dd MMM"
-                android:includeFontPadding="false"
-                android:letterSpacing="0.04"
-                android:textSize="@dimen/date_size"
-                tools:text="Thu, 30 Dec" />
-
-        </LinearLayout>
-
-        <!-- 時計とリストの間の余白 -->
-        <Space
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="0.03" />
-
-        <!-- アプリリスト + アルファベットインデックス（重ね合わせ） -->
-        <FrameLayout
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="1"
-            android:layout_marginBottom="40dp">
-
-            <!-- Niagara風 縦一列アプリリスト（カードなし、大きなアイコン） -->
-            <androidx.recyclerview.widget.RecyclerView
-                android:id="@+id/niagaraRecyclerView"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:clipChildren="false"
-                android:clipToPadding="false"
-                android:fadingEdgeLength="48dp"
-                android:paddingEnd="48dp"
-                android:paddingStart="20dp"
-                android:requiresFadingEdge="vertical" />
-
-            <!-- アルファベットインデックス（右端、波アニメーション対応） -->
-            <LinearLayout
-                android:id="@+id/appIndex"
-                android:layout_width="28dp"
-                android:layout_height="match_parent"
-                android:layout_gravity="end|center_vertical"
-                android:layout_marginEnd="2dp"
-                android:clickable="true"
-                android:focusable="true"
-                android:gravity="center"
-                android:orientation="vertical"
-                android:paddingVertical="12dp"
-                android:visibility="gone" />
-
-        </FrameLayout>
-
-    </LinearLayout>
-
-    <!-- レターポップアップ（Niagara: ドラッグ時に現在の文字を大きく表示） -->
-    <TextView
-        android:id="@+id/letterPopup"
-        android:layout_width="64dp"
-        android:layout_height="64dp"
-        android:layout_gravity="center"
-        android:alpha="0.0"
-        android:background="@drawable/letter_popup_bg"
-        android:elevation="12dp"
-        android:fontFamily="sans-serif-thin"
-        android:gravity="center"
-        android:textSize="28sp"
-        android:visibility="gone"
-        tools:text="A" />
-
-    <!-- 初回起動時のヒント -->
-    <TextView
-        android:id="@+id/firstRunTips"
-        style="@style/TextSmall"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|start"
-        android:layout_marginStart="28dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginBottom="64dp"
-        android:text="@string/swipe_up_for_apps"
-        android:visibility="gone"
-        tools:visibility="visible" />
-
-    <!-- デフォルトランチャー設定ボタン -->
-    <TextView
-        android:id="@+id/setDefaultLauncher"
-        style="@style/TextMedium"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|center_horizontal"
-        android:layout_marginHorizontal="20dp"
-        android:layout_marginBottom="64dp"
-        android:ellipsize="end"
-        android:maxLines="1"
-        android:text="@string/set_as_default_launcher_u"
-        android:visibility="gone" />
-
-    <!-- ブランド刻印（Niagara: 極小でほぼ見えない） -->
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|start"
-        android:layout_marginStart="28dp"
-        android:layout_marginBottom="24dp"
-        android:orientation="vertical"
-        android:alpha="0.25">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:fontFamily="sans-serif-light"
-            android:letterSpacing="0.1"
-            android:text="JL"
-            android:textColor="@color/qa_subtitle_color"
-            android:textSize="8sp" />
-    </LinearLayout>
-
-</FrameLayout>
