+++ app/src/main/res/layout/fragment_settings.xml
@@ -7,11 +7,12 @@
     android:animateLayoutChanges="true"
     android:background="?attr/primaryShadeColor">
 
-    <!-- BlurView for Android 11 and below (Native RenderScript Blur) -->
-    <eightbitlab.com.blurview.BlurView
+    <!-- オーバーレイ完全撤去（速度優先） -->
+    <View
         android:id="@+id/blurView"
         android:layout_width="match_parent"
-        android:layout_height="match_parent" />
+        android:layout_height="match_parent"
+        android:visibility="gone" />
 
     <ScrollView
         android:id="@+id/scrollView"
@@ -764,7 +765,7 @@
                         android:layout_gravity="end|bottom"
                         android:layout_marginEnd="4dp"
                         android:padding="8dp"
-                        android:text="@string/font_sans_thin" />
+                        android:text="@string/font_sans" />
 
                     <!-- font select layout -->
                     <LinearLayout
@@ -815,6 +816,77 @@
                             android:gravity="center"
                             android:paddingVertical="6dp"
                             android:text="@string/font_serif" />
+
+                    </LinearLayout>
+
+                </FrameLayout>
+
+                <!-- Index side selector -->
+                <FrameLayout
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:layout_marginTop="12dp"
+                    android:layout_marginBottom="4dp"
+                    android:animateLayoutChanges="true">
+
+                    <TextView
+                        style="@style/TextSmall"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:layout_marginStart="8dp"
+                        android:paddingVertical="8dp"
+                        android:text="@string/index_side"
+                        android:textColor="?attr/primaryColor" />
+
+                    <TextView
+                        android:id="@+id/indexSideText"
+                        style="@style/TextSmallBold"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:layout_gravity="end|bottom"
+                        android:layout_marginEnd="4dp"
+                        android:padding="8dp"
+                        android:text="@string/index_side_right" />
+
+                    <!-- index side select layout -->
+                    <LinearLayout
+                        android:id="@+id/indexSideSelectLayout"
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:animateLayoutChanges="true"
+                        android:background="@drawable/rounded_primary_gradient"
+                        android:elevation="8dp"
+                        android:orientation="horizontal"
+                        android:outlineAmbientShadowColor="@color/colorPrimaryDark"
+                        android:outlineSpotShadowColor="@color/colorPrimaryDark"
+                        android:paddingHorizontal="12dp"
+                        android:paddingVertical="2dp"
+                        android:visibility="gone"
+                        tools:ignore="UnusedAttribute">
+
+                        <Space
+                            android:layout_width="0dp"
+                            android:layout_height="wrap_content"
+                            android:layout_weight="1" />
+
+                        <TextView
+                            android:id="@+id/indexSideLeft"
+                            style="@style/TextSmallBold"
+                            android:layout_width="wrap_content"
+                            android:layout_height="wrap_content"
+                            android:layout_marginEnd="24dp"
+                            android:gravity="center"
+                            android:paddingVertical="6dp"
+                            android:text="@string/index_side_left" />
+
+                        <TextView
+                            android:id="@+id/indexSideRight"
+                            style="@style/TextSmallBold"
+                            android:layout_width="wrap_content"
+                            android:layout_height="wrap_content"
+                            android:gravity="center"
+                            android:paddingVertical="6dp"
+                            android:text="@string/index_side_right" />
 
                     </LinearLayout>
